fix implicit sign conversions in compat-strlcpy.c

+++ compat-strlcpy.c
@@ -52,7 +52,7 @@ strlcpy(char *dst, const char *src, size_t dsize)
 			;
 	}
 
-	return(src - osrc - 1);	/* count does not include NUL */
+	return (size_t)(src - osrc - 1);	/* count does not include NUL */
 }
 
 #endif
